Testing loading responses from config
