+++ src/GlobalServer.Tests/SettingsLoaderTests.cs
@@ -1,12 +1,10 @@
 using Xunit;
 using System;
-using System.Threading.Tasks;
-using GlobalServer.Properties;
+using GlobalServer.Tests.Files;
 using GlobalServer.Properties.Request;
-using GlobalServer.Properties.Initialization;
+using GlobalServer.Properties.Response;
 using static GlobalServer.Tests.Files.FileNames;
 using static GlobalServer.Tests.Files.FileLoader;
-using static GlobalServer.Tests.Mocks.MockFileSystemAdapter;
 
 namespace GlobalServer.Tests
 {
@@ -28,16 +26,15 @@ namespace GlobalServer.Tests
             Assert.IsType(requestType, interaction.Request);
         }
 
-        private static Task<ISettings> LoadSettings(string fileName)
+        [Theory]
+        [InlineData(RandomListResponse, typeof(RandomFromListResponse))]
+        [InlineData(FileNames.ResponseFromFile, typeof(ResponseFromFile))]
+        [InlineData(FileNames.ResponseFromString, typeof(ResponseFromString))]
+        [InlineData(FileNames.IncrementalListResponse, typeof(IncrementalListResponse))]
+        public void EachResponseCanBeLoadedFromConfig(string fileName, Type expectedType)
         {
-            const string fileLocation = @"c:\MyFile.txt";
-            var fileSystem = Create(fileLocation, GetFileContents(fileName));
-
-            var propertiesBuilder = new PropertiesBuilder()
-                .WithFileSystem(fileSystem)
-                .Build();
-
-            return propertiesBuilder.GetSettingsLoader().Load(fileLocation);
+            var settings = FileLoader.LoadSettings(fileName);
+            Assert.IsType(expectedType, settings.Result.Interactions[0].Response);
         }
     }
 }
